[FIX/#35] 소수점 값 내역 보내기, 보기 변경
1. 내역 추가, 2. 반복 내역 설정, 3. 즐겨찾기 설정

+++ domain/src/main/java/com/aos/usecase/history/PostBooksLinesUseCase.kt
@@ -6,7 +6,7 @@ import com.aos.repository.BookRepository
 class PostBooksLinesUseCase(private val bookRepository: BookRepository) {
     suspend operator fun invoke(
        bookKey: String,
-       money: Int,
+       money: Double,
        flow: String,
        asset:String,
        line: String,
